refactor: extract get_derived_classes cluster to QueryEngine (Phase 4)
Move get_derived_classes and 9 helpers from CppAnalyzer to QueryEngine:
- _check_template_param_inheritance
- _get_template_param_inheritance_indices
- _template_info_matches_name
- _build_param_name_to_index
- _resolve_param_index
- _parse_template_args
- _get_template_patterns
- _check_pattern_match
- _is_derived_from
- get_derived_classes

CppAnalyzer keeps 1-line delegation for backward compatibility.
Additional delegation wrappers added for internal methods called
by tests: _check_template_param_inheritance, _get_template_param_inheritance_indices, _parse_template_args.

Removed unused imports from CppAnalyzer: json, build_location_objects, omit_empty.

All 1435 tests pass.
